Add a library target
