+++ charon/src/generics.rs
@@ -33,7 +33,7 @@ use rustc_middle::ty::{BoundRegion, FreeRegion, PredicateKind, Region, RegionKin
 ///   is defined in an `impl` block - otherwise there are no early bound variables)
 ///   while the late bound regions are introduced by the function itself.
 ///   For example, in:
-///   ```
+///   ```text
 ///   impl<'a> Foo<'a> {
 ///       fn bar<'b>(...) -> ... { ... }
 ///   }
